third_party/blink/web_tests/storage/quota/storageinfo-request-quota-expected.txt

CONSOLE WARNING: 'window.webkitStorageInfo' is deprecated. Please use 'navigator.webkitTemporaryStorage' or 'navigator.webkitPersistentStorage' instead.
This tests webkitStorageInfo.requestQuota.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS grantedQuota is 0
PASS successfullyParsed is true

TEST COMPLETE
